Create related_items_utils.py, other refactoring
