+++ wildlifecompliance/components/call_email/models.py
@@ -9,8 +9,12 @@ from ledger.accounts.models import EmailUser, RevisionedMixin
 from ledger.licence.models import LicenceType
 from wildlifecompliance.components.main.models import CommunicationsLogEntry, UserAction, Document
 #from wildlifecompliance.components.organisations.models import Organisation
-from wildlifecompliance.components.main.models import CommunicationsLogEntry,\
-    UserAction, Document, get_related_items
+from wildlifecompliance.components.main.models import (
+        CommunicationsLogEntry,
+        UserAction, 
+        Document,
+        )
+from wildlifecompliance.components.main.related_items_utils import get_related_items
 from wildlifecompliance.components.users.models import RegionDistrict, CompliancePermissionGroup
 #from wildlifecompliance.components.main.models import InspectionType
