+++ wildlifecompliance/components/inspection/serializers.py
@@ -10,7 +10,7 @@ from wildlifecompliance.components.inspection.models import (
     InspectionCommsLogEntry,
     InspectionType,
     )
-from wildlifecompliance.components.main.models import get_related_items
+from wildlifecompliance.components.main.related_items_utils import get_related_items
 from wildlifecompliance.components.main.serializers import CommunicationLogEntrySerializer
 from wildlifecompliance.components.users.serializers import (
     ComplianceUserDetailsOptimisedSerializer,
